requirements.txt
streamlit
sentence-transformers
numpy
openai 
python-dotenv